1. Preferred metadata mappings - Added options for FanArt, BoxArt and Genre into the MetadataMappings dialog box. 2. Resolve Genre dialog, added the current genre as retrieved from the metadata source into the mapping drop down. 3. Added some validation to CopyFrontCoverFromFile and BackCoverFromFile.

+++ trunk/OMLEngine/Title.cs
@@ -2151,34 +2151,42 @@ namespace OMLEngine
         // copy front cover image and set the menu cover too (resized version)
         public bool CopyFrontCoverFromFile(string source, bool deleteSource)
         {
-            try
+            if (!string.IsNullOrEmpty(source))
             {
-                File.Copy(source, GetDefaultFrontCoverName(), true);
-                if (deleteSource) File.Delete(source);
-                FrontCoverPath = GetDefaultFrontCoverName();
-                BuildResizedMenuImage();
-                return true;
+                try
+                {
+                    File.Copy(source, GetDefaultFrontCoverName(), true);
+                    if (deleteSource) File.Delete(source);
+                    FrontCoverPath = GetDefaultFrontCoverName();
+                    BuildResizedMenuImage();
+                    return true;
+                }
+                catch
+                {
+                    return false;
+                }
             }
-            catch
-            {
-                return false;
-            }
+            return false;
         }
 
         public bool CopyBackCoverFromFile(string source, bool deleteSource)
         {
-            try
+            if (!string.IsNullOrEmpty(source))
             {
-                File.Copy(source, GetDefaultBackCoverName(), true);
-                if (deleteSource) File.Delete(source);
-                BackCoverPath = GetDefaultBackCoverName();
-                
-                return true;
+                try
+                {
+                    File.Copy(source, GetDefaultBackCoverName(), true);
+                    if (deleteSource) File.Delete(source);
+                    BackCoverPath = GetDefaultBackCoverName();
+
+                    return true;
+                }
+                catch
+                {
+                    return false;
+                }
             }
-            catch
-            {
-                return false;
-            }
+            return false;
         }
 
         public string GetDefaultNoCoverName()
